Split relevant DecodeUtils functions into coil-gif and coil-svg.
